Batch File Processing
This project will have teams building a batch file processor. The scope
of this effort is to build a tool that parses and executes a batch file
containing a number of commands. A batch file, or batch, contains one or
more commands that are executed sequentially. Each command executed by
our batch processor will be executed as a process and communicate (pass
information) using files or pipes.