Testing Code
This project will produce library JAR file containing a TaskExecutor
service. TaskExecutor is a service class that maintains a pool
(collection) of N threads that are used to execute instances of Tasks
provided by the TaskExecutor‟s clients.
The primary goal of this project is to have team implement the
multithreaded synchronization needed to implement the TaskExecutor and
internal Blocking FIFO without the use of Java's concurrency library
(other than Semaphore) and without synchronized methods.