Data Store Client – Server Project
This project will involve constructing a client-server system that
communicates over TCP/IP sockets. The server will store, retrieve,
delete, & list named data sets on the server. Each data set is
associated with a string name provided in the client’srequest to the
server. The server is responsible for storing this data in a persistent
fashion for later retrieval. Data is persistently stored in files in a
directory under the project directory.